# docs/api/solem-api.openapi.yaml
openapi: 3.1.0
info:
  title: SOLEM API
  version: 0.1.0-step0
  description: |
    API di sistema di SOLEM — OS AI-native che ospita GAVIO e (futuro) altre AI.
    Progettata per essere consumata da AI come consumer primario.

    Filosofia: schema strutturato ovunque, errori machine-readable, output
    deterministici, no stringhe libere non parseable.
  contact:
    name: SOLEM
    email: r.guidotti@theoryholding.com
servers:
  - url: http://localhost:8001
    description: VM locale (default)
  - url: http://solem.local:8001
    description: mDNS bare-metal (Step 1+)
...
paths:
...
  /solem/pairing/confirm:
    post:
      tags: [pairing]
      operationId: pairing_confirm
      summary: Completa pairing — il device riceve config mesh + cert
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PairingConfirmRequest' }
      responses:
        '200':
          description: Pairing completato — risposta contiene WG config + opzionale cert mTLS
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PairingConfirmResponse' }
        '404':
          description: PIN sconosciuto o già consumato
        '410':
          description: PIN scaduto
...
components:
  schemas:
...
    PairingConfirmRequest:
      type: object
      required: [pin, device_name, device_pubkey_wg]
      properties:
        pin: { type: string, minLength: 8, maxLength: 8 }
        device_name: { type: string, minLength: 1, maxLength: 64 }
        device_pubkey_wg:
          type: string
          description: WireGuard public key (base64)
        device_pubkey_mtls:
          type: string
          description: CSR PEM per cert mTLS (opzionale)
    PairingConfirmResponse:
      type: object
      required: [device_id, assigned_ip, coordinator_pubkey_wg, mesh_subnet, dns_server]
      properties:
        device_id: { type: string }
        assigned_ip:
          type: string
          example: 10.42.0.10/32
        coordinator_pubkey_wg: { type: string }
        ca_cert_pem: { type: string, nullable: true }
        client_cert_pem: { type: string, nullable: true }
        mesh_subnet: { type: string, default: "10.42.0.0/24" }
        dns_server: { type: string, default: "10.42.0.1" }